Use universal container with Java (for TestRail)
Now using the large, multi-language universal GitHub Codespaces
container[1], as (in addition to Python) we also need Java (for the
Gauge TestRail plugin).

This commit just brings in the container without any modifications,
meaning we will need to add Gauge to the container setup in a subsequent
commit, just as we did in a previous commit
(c33490049a0e0a666e0903ef327b98a4f79294d5) for the Python container that
we have been using until now.

[1]: https://github.com/microsoft/vscode-dev-containers/tree/master/containers/codespaces-linux